Nevermind, move it to a separate file
